HoK_RolePlay/requirements.txt
numpy==1.23.4
face_alignment==1.3.5
imageio==2.19.3
imageio-ffmpeg==0.4.7
librosa==0.9.2 
numba
zhconv
resampy==0.3.1
pydub==0.25.1 
scipy==1.10.1
kornia==0.6.8
tqdm
yacs==0.1.8
pyyaml  
joblib==1.1.0
facexlib==0.3.0
gradio==4.16.0
edge-tts>=6.1.9
openai-whisper
scikit-image==0.19.3
accelerate
transformers==4.32.0
einops 
transformers_stream_generator==0.0.4
sentencepiece
google-generativeai
tiktoken
accelerate
protobuf==3.19.6
openai
google-api-python-client==2.126.0
g4f
librosa==0.9.2
numba==0.56.4
pytorch-lightning
ffmpeg-python
onnxruntime
tqdm
funasr>=1.0.0
cn2an
pypinyin
pyopenjtalk
g2p_en
modelscope==1.10.0
sentencepiece
chardet
PyYAML
psutil
jieba_fast
jieba
LangSegment
# gfpgan
# ========Qwen Need========#
# transformers==4.32.0
# accelerate 
# tiktoken
# scipy
# transformers_stream_generator==0.0.4 
# peft 
# deepspeed
